Agregados cambios a ABM Tarjeta

+++ TP_PAVI/Entidades/Tarjeta.cs
@@ -10,7 +10,7 @@ namespace AppBTS.Entidades
     public class Tarjeta
     {
         public Marca_tarjeta Id_marca_tarjeta { get; set; }
-        public int Nro_tarjeta { get; set;}
+        public long Nro_tarjeta { get; set;}
         public DateTime Fecha_venc { get; set; }
         public Usuario Id_usuario { get; set; }
         public bool Borrado { get; set; }
